pyproject.toml
[project]
name = "multi-tts-runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.18.0",
    "numpy>=2.2.6",
    "openai>=1.82.1",
    "scipy>=1.15.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]

[tool.ruff]
line-length = 120
